Added UserController, changed HTTP-methods
